Clear IDB object store after a file is downloaded
Previously, chunks of the current file were removed after being accessed, but this still left the possibility of data being left in the object store if the page was closed after decrypting but before downloading. This is still possible, but the addition of clearing the objectstore after a successful download means that the issue will resolve itself by simply using transpo an additional time.